% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_funs.R
\name{add_holdout_ind}
\alias{add_holdout_ind}
\title{Define and fit growth models evaluated on holdout observations.}
\usage{
add_holdout_ind(data, ID, hold_column = "hold", random = TRUE,
  seed = NULL)
}
\arguments{
\item{data}{Input dataset, can be a \code{data.frame} or a \code{data.table}.}

\item{ID}{A character string name of the column that contains the unique subject identifiers.}

\item{hold_column}{A name of the column that will contain the the holdout indicators}

\item{random}{Logical, specifying if the holdout observations should be selected at random.
If FALSE then the last observation for each subject is selected as a holdout.}

\item{seed}{Random number seed for selecting a random holdout.}
}
\value{
An input data with added holdout indicator column (TRUE for holdout observation indicator, FALSE for training observation indicator).
}
\description{
Define and fit growth models evaluated on holdout observations.
}
