% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{as.int}
\alias{as.int}
\title{Convert specific columns in \code{vars} to numeric}
\usage{
as.int(data, vars)
}
\arguments{
\item{data}{Input dataset, as \code{data.table}.}

\item{vars}{Column name(s) that should be converted to numeric type.}
}
\description{
Convert specific columns in \code{vars} to numeric
}
