% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{char_to_factor}
\alias{char_to_factor}
\title{Convert all character columns to factors}
\usage{
char_to_factor(data, skip_vars)
}
\arguments{
\item{data}{Input dataset, as \code{data.table}.}

\item{skip_vars}{These columns will not be converted to factor}
}
\description{
Convert all character columns to factors
}
