% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hbgd_fittedTrajectory.R
\name{create_all_fittedTrajectory}
\alias{create_all_fittedTrajectory}
\title{Create a fittedTrajectory object with fits for a single subject}
\usage{
create_all_fittedTrajectory(subj_dat, grid_fits_dat, holdout_fits_dat = NULL,
  ID_var, t_var, y_var, sex_var, method, xy_pair_name = c("agedays", "htcm"),
  grid_fits_var = "SL.preds", holdout_fits_var = "holdout_fit",
  checkpoints = c(50, 100), fun_y_to_raw = hbgd::who_zscore2htcm,
  fun_y_to_z = function(x, y, ...) return(y))
}
\arguments{
\item{subj_dat}{...}

\item{grid_fits_dat}{...}

\item{holdout_fits_dat}{...}

\item{ID_var}{A character string name of the column in \code{grid_fits_dat} that contains the unique subject identifiers.}

\item{t_var}{A character string name of the column with integer-valued measurement time-points (in days, weeks, months, etc).}

\item{y_var}{A character string name of the column that represent the response variable in the model.}

\item{sex_var}{...}

\item{method}{...}

\item{xy_pair_name}{...}

\item{grid_fits_var}{...}

\item{holdout_fits_var}{...}

\item{fun_y_to_raw}{...}

\item{fun_y_to_z}{...}

\item{ID}{The value of the subject ID for which the trajectory object will be created.}
}
\value{
...
}
\description{
Creates objects of class \code{fittedTrajectory} with growth curve predictions, one object for subject ID in the data.
This can be used for further growth curve analysis with hbgd R package.
Requires the the dataset containing the predicted growth curves for all subjects \code{grid_fits_dat}.
}

