% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelPredictionModel.R
\name{create_fit_object}
\alias{create_fit_object}
\title{Create a model fit list}
\usage{
create_fit_object(model.fit, model_alg, fitfunname, params, coef, nobs,
  model_contrl, fitclass = "H2Omodel", ...)
}
\arguments{
\item{model.fit}{Model fit object}

\item{model_alg}{Name of the model algorithm}

\item{fitfunname}{The name of the main function used for fitting (e.g., "glm")}

\item{params}{List of main model parameters}

\item{coef}{Fitted glm coefficients, if used}

\item{nobs}{Number of observations used for fitting the model}

\item{model_contrl}{Full list of model control parameters used}

\item{fitclass}{The name of the R6 object class used for fitting}

\item{...}{Additional objects that will be included in the final fit list}
}
\description{
Create a model fit list
}
