% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelPredictionModel.R
\name{create_fit_params}
\alias{create_fit_params}
\title{Create a list with main model parameters}
\usage{
create_fit_params(reg)
}
\arguments{
\item{reg}{RegressionClass Object}
}
\description{
Create a list with main model parameters
}
