% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hbgd_fittedTrajectory.R
\name{create_fittedTrajectory_1subj}
\alias{create_fittedTrajectory_1subj}
\title{Create a fittedTrajectory object with fits for a single subject}
\usage{
create_fittedTrajectory_1subj(subj_dat_1subj, grid_fits_dat_1subj,
  holdout_fits_dat_1subj = NULL, t_var, y_var, sex_var, method,
  xy_pair_name = c("agedays", "htcm"), grid_fits_var = "SL.preds",
  holdout_fits_var = "holdout_fit", checkpoints = c(50, 100),
  fun_y_to_raw = hbgd::who_zscore2htcm, fun_y_to_z = function(x, y, ...)
  return(y))
}
\arguments{
\item{subj_dat_1subj}{Data on a single subject}

\item{grid_fits_dat_1subj}{Grid of equally spaced model predictions for a single subject}

\item{holdout_fits_dat_1subj}{Data with model predictions for holdout datapoints for a single subject (optional)}

\item{t_var}{A character name of the column in \code{subj_dat}, \code{grid_fits_dat} and \code{holdout_fits_dat} with integer-valued time measurements (e.g., days, weeks, months).}

\item{y_var}{A character name of the column that represent the response variable in the model (the outcome).}

\item{sex_var}{A character name of the column that represent the response variable in the model (the outcome).}

\item{method}{A name of the modeling approach for future model comparison.}

\item{xy_pair_name}{... Same as in hbgd package...}

\item{grid_fits_var}{The column name in \code{grid_fits_dat} that contains the modeling predictions for each subject-time row.}

\item{holdout_fits_var}{The column name in \code{holdout_fits_dat} that contains the modeling predictions for each holdout subject-time row.}

\item{fun_y_to_raw}{The function to convert the scale of the observed outcome and predictions into the raw scale (height / weight).
If the fits are already on the raw scale then specify this argument as \code{function(x, y, ...) return(y)}.}

\item{fun_y_to_z}{The function to convert the scale of the observed outcome and predictions into the z-scale.
If the fits are already on the z-scale then specify this argument as \code{function(x, y, ...) return(y)}.}
}
\value{
An object of class \code{fittedTrajectory} for further downstream analysis.
}
\description{
Creates an object of class \code{fittedTrajectory} with growth curve predictions.
This can be used for further growth curve analysis with hbgd R package.
Requires the the dataset containing the predicted growth curves for a single subject \code{grid_fits_dat_1subj}.
}

