% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthcurve_features.R
\name{define_features}
\alias{define_features}
\title{Define features (predictors) for training or testing (validation) data}
\usage{
define_features(dataDT, nodes, train_set = TRUE, holdout = TRUE,
  hold_column = "hold", noNAs = FALSE, includeRLMIDind = FALSE,
  verbose = getOption("longDiSL.verbose"))
}
\arguments{
\item{dataDT}{Input data.table}

\item{nodes}{...}

\item{train_set}{...}

\item{holdout}{...}

\item{hold_column}{...}

\item{noNAs}{...}

\item{includeRLMIDind}{...}

\item{verbose}{Set to \code{TRUE} to print messages on status and information to the console. Turn this on by default using \code{options(longDiSL.verbose=TRUE)}.}
}
\value{
...
}
\description{
Define features (predictors) for training or testing (validation) data
}

