% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_estimation.R
\name{define_LR_summaries}
\alias{define_LR_summaries}
\title{Define various summaries of observed outcome}
\usage{
define_LR_summaries(OData, holdout = FALSE,
  verbose = getOption("growthcurveSL.verbose"))
}
\arguments{
\item{OData}{Input data object created by \code{importData} function.}

\item{holdout}{...}

\item{verbose}{Set to \code{TRUE} to print messages on status and information to the console. Turn this on by default using \code{options(growthcurveSL.verbose=TRUE)}.}
}
\value{
...
}
\description{
Define various summaries of observed outcome
}

