% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_estimation.R
\name{define_predictors}
\alias{define_predictors}
\title{Define predictors for training or testing (validation) data}
\usage{
define_predictors(dataDT, nodes, train_set = TRUE, holdout = TRUE,
  hold_column = "hold", verbose = getOption("growthcurveSL.verbose"))
}
\arguments{
\item{dataDT}{Input data.table}

\item{nodes}{...}

\item{train_set}{...}

\item{holdout}{...}

\item{hold_column}{...}

\item{verbose}{Set to \code{TRUE} to print messages on status and information to the console. Turn this on by default using \code{options(growthcurveSL.verbose=TRUE)}.}
}
\value{
...
}
\description{
Define predictors for training or testing (validation) data
}

