% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_main.R
\name{defLearner}
\alias{+.ModelStack}
\alias{defLearner}
\title{Define modeling algorithm(s), package and parameters}
\usage{
defLearner(estimator, ...)

\method{+}{ModelStack}(sVar1, sVar2)
}
\arguments{
\item{estimator}{A character string name of package and estimator (algorithm) name, separated by "_".}

\item{...}{Additional modeling parameters to be passed to modeling function.}

\item{sVar1}{An object returned by a call to \code{defLearner} or \code{defGrid} functions.}

\item{sVar2}{An object returned by a call to \code{defLearner} or \code{defGrid} functions.}
}
\description{
Define modeling algorithm(s), package and parameters
}

