% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{drop_NA_y}
\alias{drop_NA_y}
\title{Drop all observation rows with missing outcomes}
\usage{
drop_NA_y(data, OUTCOME)
}
\arguments{
\item{data}{Input dataset, as \code{data.table}.}

\item{OUTCOME}{Character name of the column of outcomes.}
}
\description{
Drop all observation rows with missing outcomes
}
