% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelingSL_main.R
\name{eval_MSE_CV}
\alias{eval_MSE_CV}
\title{Evaluate cross-validated MSE}
\usage{
eval_MSE_CV(modelfit, newdata, verbose = getOption("growthcurveSL.verbose"))
}
\arguments{
\item{modelfit}{Model fit object returned by \code{\link{fit_model}} function.}

\item{newdata}{...}

\item{verbose}{Set to \code{TRUE} to print messages on status and information to the console. Turn this on by default using \code{options(growthcurveSL.verbose=TRUE)}.}
}
\value{
...
}
\description{
Evaluate cross-validated MSE
}

