% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{factor_to_dummy}
\alias{factor_to_dummy}
\title{Convert factors to binary indicators, for factors with > 2 levels drop the first factor level and define several dummy variables for the rest of the levels}
\usage{
factor_to_dummy(data, skip_vars)
}
\arguments{
\item{data}{Input dataset, as \code{data.table}.}

\item{skip_vars}{These columns will not be converted}
}
\description{
Convert factors to binary indicators, for factors with > 2 levels drop the first factor level and define several dummy variables for the rest of the levels
}

