% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_SuperLearner_funs.R
\name{fit_cvSL}
\alias{fit_cvSL}
\title{Discrete SuperLearner with V-fold cross-validation.}
\usage{
fit_cvSL(ID, t_name, x, y, data, models, nfolds = 5, fold_column = NULL,
  refit = TRUE, seed = NULL, verbose = getOption("GriDiSL.verbose"), ...)
}
\arguments{
\item{ID}{A character string name of the column that contains the unique subject identifiers.}

\item{t_name}{A character string name of the column with integer-valued measurement time-points (in days, weeks, months, etc).}

\item{x}{A vector containing the names of predictor variables to use for modeling. If x is missing, then all columns except \code{ID}, \code{y} are used.}

\item{y}{A character string name of the column that represent the response variable in the model.}

\item{data}{Input dataset, can be a \code{data.frame} or a \code{data.table}.}

\item{models}{Parameters specifying the type of modeling procedure to be used.}

\item{nfolds}{Number of folds to use in cross-validation.}

\item{fold_column}{The name of the column in the input data that contains the cross-validation fold indicators (must be an ordered factor).}

\item{refit}{Set to \code{TRUE} (default) to refit the best estimator using the entire dataset.
When \code{FALSE}, it might be impossible to make predictions from this model fit.}

\item{seed}{Random number seed for selecting a random holdout.}

\item{verbose}{Set to \code{TRUE} to print messages on status and information to the console. Turn this on by default using \code{options(GriDiSL.verbose=TRUE)}.}
}
\value{
...
}
\description{
Define and fit discrete SuperLearner for growth curve modeling.
Model selection (scoring) is based on V-fold cross-validated MSE that leaves entire subjects outside of the training sample.
This is in contrast to holdout SuperLearner in \code{\link{fit_holdoutSL}} that leaves only a single random (or last) growth measurement  outside of the training sample.
}

