% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthcurve_model.R
\name{fit_growthcurveSL}
\alias{fit_growthcurveSL}
\title{Define and fit growth models with cross-validation.}
\usage{
fit_growthcurveSL(ID, t_name, x, y, data, params, nfolds = 5,
  fold_column = NULL, seed = NULL, expr_to_train = NULL,
  use_new_features = FALSE, verbose = getOption("growthcurveSL.verbose"))
}
\arguments{
\item{ID}{A character string name of the column that contains the unique subject identifiers.}

\item{t_name}{A character string name of the column with integer-valued measurement time-points (in days, weeks, months, etc).}

\item{x}{A vector containing the names of predictor variables to use for modeling. If x is missing, then all columns except \code{ID}, \code{y} are used.}

\item{y}{A character string name of the column that represent the response variable in the model.}

\item{data}{Input dataset, can be a \code{data.frame} or a \code{data.table}.}

\item{params}{Parameters specifying the type of modeling procedure to be used.}

\item{nfolds}{Number of folds to use in cross-validation.}

\item{fold_column}{The name of the column in the input data that contains the cross-validation fold indicators (must be an ordered factor).}

\item{seed}{Random number seed for selecting a random holdout.}

\item{expr_to_train}{Additional logical expression which will further subset observations (rows) for training data.
Use this to restrict the model fitting to a specific subsample of the training datset.}

\item{use_new_features}{...}

\item{verbose}{Set to \code{TRUE} to print messages on status and information to the console. Turn this on by default using \code{options(growthcurveSL.verbose=TRUE)}.}
}
\value{
...
}
\description{
Define and fit growth models with cross-validation.
}

