% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthcurve_model.R
\name{fit_holdoutSL}
\alias{fit_holdoutSL}
\title{Growth curve SuperLearner with one-out holdout validation}
\usage{
fit_holdoutSL(ID, t_name, x, y, data, params, hold_column = NULL,
  random = FALSE, seed = NULL, expr_to_train = NULL,
  use_new_features = FALSE, verbose = getOption("longDiSL.verbose"))
}
\arguments{
\item{ID}{A character string name of the column that contains the unique subject identifiers.}

\item{t_name}{A character string name of the column with integer-valued measurement time-points (in days, weeks, months, etc).}

\item{x}{A vector containing the names of predictor variables to use for modeling. If x is missing, then all columns except \code{ID}, \code{y} are used.}

\item{y}{A character string name of the column that represent the response variable in the model.}

\item{data}{Input dataset, can be a \code{data.frame} or a \code{data.table}.}

\item{params}{Parameters specifying the type of modeling procedure to be used.}

\item{hold_column}{The name of the column that contains the holdout observation indicators (TRUE/FALSE) in the input data.
This holdout column must be defined and added to the input data prior to calling this function.}

\item{random}{Logical, specifying if the holdout observations should be selected at random.
If FALSE then the last observation for each subject is selected as a holdout.}

\item{seed}{Random number seed for selecting a random holdout.}

\item{expr_to_train}{Additional logical expression which will further subset observations (rows) for training data.
Use this to restrict the model fitting to a specific subsample of the training datset.}

\item{use_new_features}{...}

\item{verbose}{Set to \code{TRUE} to print messages on status and information to the console. Turn this on by default using \code{options(longDiSL.verbose=TRUE)}.}
}
\value{
...
}
\description{
Define and fit discrete SuperLearner for growth curve modeling.
Model selection (scoring) is based on MSE for a single random (or last) holdout data-point for each subject.
This is in contrast to the model selection with V-fold cross-validated MSE in \code{\link{fit_cvSL}},
which leaves the entire subjects (entire growth curves) outside of the training sample.
}

