% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelingSL_main.R
\name{fit_model}
\alias{fit_model}
\title{Generic modeling function for any longitudinal data.}
\usage{
fit_model(ID, t_name, x, y, train_data, valid_data, params, nfolds, fold_column,
  seed, useH2Oframe = FALSE, subset_exprs = NULL, subset_idx = NULL,
  verbose = getOption("longDiSL.verbose"))
}
\arguments{
\item{ID}{A character string name of the column that contains the unique subject identifiers.}

\item{t_name}{A character string name of the column with integer-valued measurement time-points (in days, weeks, months, etc).}

\item{x}{A vector containing the names of predictor variables to use for modeling. If x is missing, then all columns except \code{ID}, \code{y} are used.}

\item{y}{A character string name of the column that represent the response variable in the model.}

\item{train_data}{Input dataset, can be a \code{data.frame} or a \code{data.table}.}

\item{valid_data}{Optional \code{data.frame} or \code{data.table} with validation data. When provided, this dataset will be used for scoring the model fit(s).}

\item{params}{Parameters specifying the type of modeling procedure to be used.}

\item{nfolds}{...}

\item{fold_column}{...}

\item{seed}{...}

\item{useH2Oframe}{...}

\item{subset_exprs}{...}

\item{subset_idx}{...}

\item{verbose}{Set to \code{TRUE} to print messages on status and information to the console. Turn this on by default using \code{options(longDiSL.verbose=TRUE)}.}
}
\value{
...
}
\description{
Generic modeling function for any longitudinal data.
}

