% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_estimation.R
\name{get_fit}
\alias{get_fit}
\title{Define and fit growth models.}
\usage{
get_fit(OData, OData_train, OData_valid, predvars, params, holdout = TRUE,
  hold_column = NULL, random = FALSE, seed = NULL, expr_to_train = NULL,
  verbose = getOption("growthcurveSL.verbose"))
}
\arguments{
\item{OData}{Input data object created by \code{importData} function.}

\item{OData_train}{...}

\item{OData_valid}{...}

\item{predvars}{...}

\item{params}{...}

\item{holdout}{...}

\item{hold_column}{..}

\item{random}{...}

\item{seed}{...}

\item{verbose}{Set to \code{TRUE} to print messages on status and information to the console. Turn this on by default using \code{options(growthcurveSL.verbose=TRUE)}.}
}
\value{
...
}
\description{
Define and fit growth models.
}

