% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_SuperLearner_funs.R
\name{get_out_of_sample_predictions}
\alias{get_out_of_sample_predictions}
\title{Get the combined out of sample predictions from V cross-validation models}
\usage{
get_out_of_sample_predictions(modelfit)
}
\arguments{
\item{modelfit}{A model object of class \code{PredictionModel} returned by functions \code{fit_model} or \code{fit}.}
}
\value{
A vector of out-of-sample predictions from the best selected model (CV-MSE).
}
\description{
Get the combined out of sample predictions from V cross-validation models
}
