% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SuperLearner_main.R
\name{get_validation_data}
\alias{get_validation_data}
\title{Get validation data used by the modeling object}
\usage{
get_validation_data(modelfit)
}
\arguments{
\item{modelfit}{A model object of class \code{PredictionModel} returned by functions \code{fit_model}, \code{fit_holdoutSL} or \code{fit_cvSL}.}
}
\value{
\code{data.table} that was used for model scoring (CV-MSE).
}
\description{
Wrapper function for obtaining the validation dataset saved in the modeling object.
}

