% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_SuperLearner_funs.R
\name{get_yvalues}
\alias{get_yvalues}
\title{Get the y values (outcomes) used in the training data}
\usage{
get_yvalues(modelfit, subset_idx = NULL, ...)
}
\arguments{
\item{modelfit}{A model object of class \code{PredictionModel} returned by functions \code{fit_model} or \code{fit}.}

\item{subset_idx}{The index of rows (subset of training data) that was actually used for buiding / fitting the models}
}
\value{
A vector of out-of-sample predictions from the best selected model (CV-MSE).
}
\description{
Get the y values (outcomes) used in the training data
}
