% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{growthcurveSLOptions}
\alias{growthcurveSLOptions}
\title{Querying/setting a single \code{growthcurveSL} option}
\usage{
growthcurveSLOptions(o, value)
}
\arguments{
\item{o}{Option name (string). See \code{\link{set_all_growthcurveSL_options}}.}

\item{value}{Value to assign (optional)}
}
\description{
To list all \code{growthcurveSL} options, just run this function without any parameters provided. To query only one value, pass the first parameter. To set that, use the \code{value} parameter too.
}
\details{
The arguments of \code{\link{set_all_growthcurveSL_options}} list all available \code{growthcurveSL} options.
}
\examples{
\dontrun{
growthcurveSLOptions()
growthcurveSLOptions('fit.package')
growthcurveSLOptions('fit.package', 'h2o')
}
}
\seealso{
\code{\link{set_all_growthcurveSL_options}}
}

