% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_estimation.R
\name{importData}
\alias{importData}
\title{Import data, define nodes (columns), define dummies for factor columns and define input data R6 object}
\usage{
importData(data, ID = "Subject_ID", t_name = "time_period", covars,
  OUTCOME = "Y", verbose = getOption("growthcurveSL.verbose"))
}
\arguments{
\item{data}{Input dataset, can be a data.frame or a data.table}

\item{ID}{...}

\item{t_name}{...}

\item{covars}{...}

\item{OUTCOME}{...}

\item{verbose}{Set to \code{TRUE} to print messages on status and information to the console. Turn this on by default using \code{options(growthcurveSL.verbose=TRUE)}.}

\item{MONITOR}{...}
}
\value{
...
}
\description{
Import data, define nodes (columns), define dummies for factor columns and define input data R6 object
}

