% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{logical_to_int}
\alias{logical_to_int}
\title{Convert logical covariates to integers}
\usage{
logical_to_int(data, skip_vars)
}
\arguments{
\item{data}{Input dataset, as \code{data.table}.}

\item{skip_vars}{These columns will not be converted to integer}
}
\description{
Convert logical covariates to integers
}
