% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{longDiSLOptions}
\alias{longDiSLOptions}
\title{Querying/setting a single \code{longDiSL} option}
\usage{
longDiSLOptions(o, value)
}
\arguments{
\item{o}{Option name (string). See \code{\link{set_all_longDiSL_options}}.}

\item{value}{Value to assign (optional)}
}
\description{
To list all \code{longDiSL} options, just run this function without any parameters provided. To query only one value, pass the first parameter. To set that, use the \code{value} parameter too.
}
\details{
The arguments of \code{\link{set_all_longDiSL_options}} list all available \code{longDiSL} options.
}
\examples{
\dontrun{
longDiSLOptions()
longDiSLOptions('fit.package')
longDiSLOptions('fit.package', 'h2o')
}
}
\seealso{
\code{\link{set_all_longDiSL_options}}
}

