% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{longGriDiSLOptions}
\alias{longGriDiSLOptions}
\title{Querying/setting a single \code{longGriDiSL} option}
\usage{
longGriDiSLOptions(o, value)
}
\arguments{
\item{o}{Option name (string). See \code{\link{set_all_longGriDiSL_options}}.}

\item{value}{Value to assign (optional)}
}
\description{
To list all \code{longGriDiSL} options, just run this function without any parameters provided. To query only one value, pass the first parameter. To set that, use the \code{value} parameter too.
}
\details{
The arguments of \code{\link{set_all_longGriDiSL_options}} list all available \code{longGriDiSL} options.
}
\examples{
\dontrun{
longGriDiSLOptions()
longGriDiSLOptions('fit.package')
longGriDiSLOptions('fit.package', 'h2o')
}
}
\seealso{
\code{\link{set_all_longGriDiSL_options}}
}

