% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_funs.R
\name{make_kfold_from_column}
\alias{make_kfold_from_column}
\title{Convert a column of validation folds into origami format}
\usage{
make_kfold_from_column(fold_column)
}
\arguments{
\item{fold_column}{A vector with validation fold indicators}
}
\value{
An object (list) or validation / training folds of length V, for V-fold cross-validation
}
\description{
From a column of validation folds (length n) into an origami-based format for fold storage
for V-fold cross-validation
}
