% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{make_model_report}
\alias{make_model_report}
\title{Generate report(s) with modeling stats using pandoc.}
\usage{
make_model_report(modelfit, data, K = 5, format = c("html", "pdf", "word"),
  file.name = getOption("gridisl.file.name"),
  file.path = getOption("gridisl.file.path"), openFile = TRUE,
  keep_md = FALSE, keep_tex = FALSE, ...)
}
\arguments{
\item{modelfit}{Model fit object returned by model fitting functions.}

\item{data}{Input dataset used for model fitting (optional).}

\item{K}{The number of top performing models for which to provide performance measures. Defaults to 5 or the total number of models
fit, whichever is smaller.}

\item{format}{Choose the Pandoc output format for the report file (html, pdf or word).
Note that the html report file is always produced in addition to any other selected format.}

\item{file.name}{File name for the report file without extension. Default file name is assigned based on the current date.}

\item{file.path}{Directory path where the report file(s) should be written. Default is to use the system temporary directory.}

\item{openFile}{Open the report file with OS default viewer?}

\item{keep_md}{Keep the source .md files?}

\item{keep_tex}{Keep the source .tex files for pdf output?}

\item{...}{Additional arguments may specify the report title (\code{author}), author (\code{title}).
Specifying the logical flag \code{only.coefs=TRUE} disables printing of all h2o-specific model summaries.}
}
\value{
String specifying the path to the main report file.
}
\description{
Generate report(s) with modeling stats using pandoc.
}
