% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelPredictionStack.R
\name{make_PredictionStack}
\alias{make_PredictionStack}
\title{Combine models into ensemble}
\usage{
make_PredictionStack(...)
}
\arguments{
\item{...}{Different objects of class "PredictionModel" separated by a comma.}
}
\description{
Combine several fitted models into a single ensemble model of class 'PredictionStack'.
}
