% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthcurve_model.R
\name{predict_fullcurve}
\alias{predict_fullcurve}
\title{Predict for a holdout set (NOT IMPLEMENTED)}
\usage{
predict_fullcurve(OData, tmin, tmax, modelIDs,
  verbose = getOption("growthcurveSL.verbose"))
}
\arguments{
\item{OData}{Input data object created by \code{importData} function.}

\item{tmin}{...}

\item{tmax}{...}

\item{modelIDs}{...}

\item{verbose}{Set to \code{TRUE} to print messages on status and information to the console. Turn this on by default using \code{options(growthcurveSL.verbose=TRUE)}.}
}
\value{
...
}
\description{
Predict for a holdout set (NOT IMPLEMENTED)
}

