% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthcurve_model.R
\name{predict_growthcurve_holdout}
\alias{predict_growthcurve_holdout}
\title{Predict for a holdout set}
\usage{
predict_growthcurve_holdout(modelfit, valid_data,
  predict_only_best_model = TRUE,
  verbose = getOption("growthcurveSL.verbose"))
}
\arguments{
\item{modelfit}{Model fit object returned by \code{\link{get_fit}} function.}

\item{valid_data}{...}

\item{predict_only_best_model}{... Not implemented ...}

\item{verbose}{Set to \code{TRUE} to print messages on status and information to the console. Turn this on by default using \code{options(growthcurveSL.verbose=TRUE)}.}
}
\value{
...
}
\description{
Predict for a holdout set
}

