% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthcurve_model.R
\name{predict_growthcurveSL}
\alias{predict_growthcurveSL}
\title{Predict for cross-validation sets}
\usage{
predict_growthcurveSL(modelfit, valid_data, predict_only_best_model = TRUE,
  verbose = getOption("growthcurveSL.verbose"))
}
\arguments{
\item{modelfit}{Model fit object returned by \code{\link{get_fit}} function.}

\item{valid_data}{...}

\item{predict_only_best_model}{... Not implemented ...}

\item{verbose}{Set to \code{TRUE} to print messages on status and information to the console. Turn this on by default using \code{options(growthcurveSL.verbose=TRUE)}.}
}
\value{
...
}
\description{
Predict for cross-validation sets
}

