% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelingSL_main.R
\name{predict_holdout}
\alias{predict_holdout}
\title{Prediction for holdout (out-of-sample) model}
\usage{
predict_holdout(modelfit, newdata, predict_only_bestK_models,
  add_subject_data = FALSE, subset_idx = NULL,
  verbose = getOption("growthcurveSL.verbose"))
}
\arguments{
\item{modelfit}{Model fit object returned by \code{\link{fit_holdoutSL}} or \code{\link{fit_cvSL}} functions.}

\item{newdata}{Subject-specific data for which holdout (out-of-sample) predictions should be obtained.}

\item{predict_only_bestK_models}{Specify the total number of top-ranked models (validation or C.V. MSE) for which predictions should be obtained.
Leave missing to obtain predictions for all models that were fit as part of this ensemble.}

\item{add_subject_data}{Set to \code{TRUE} to add the subject-level data to the resulting predictions (returned as a data.table).
When \code{FALSE} (default) only the actual predictions are returned (as a matrix with each column representing predictions from a specific model).}

\item{subset_idx}{...}

\item{verbose}{Set to \code{TRUE} to print messages on status and information to the console. Turn this on by default using \code{options(growthcurveSL.verbose=TRUE)}.}
}
\value{
...
}
\description{
When \code{newdata} is missing there are two possible types of holdout predictions, depending on the modeling approach.
1. For \code{fit_holdoutSL} the default holdout predictions will be based on validation data.
2. For \code{fit_cvSL} the default is to leave use the previous out-of-sample (holdout) predictions from the training data.
}

