% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthcurve_model.R
\name{predict_holdouts_only}
\alias{predict_holdouts_only}
\title{Predict for holdout observations only}
\usage{
predict_holdouts_only(modelfit, predict_only_bestK_models,
  add_subject_data = FALSE, verbose = getOption("growthcurveSL.verbose"))
}
\arguments{
\item{modelfit}{Model fit object returned by \code{\link{get_fit}} function.}

\item{predict_only_bestK_models}{Specify the total number of top-ranked models (validation or C.V. MSE) for which predictions should be obtained.
Leave missing to obtain predictions for all models that were fit as part of this ensemble.}

\item{add_subject_data}{Set to \code{TRUE} to add the subject-level data to the resulting predictions (returned as a data.table).
When \code{FALSE} (default) only the actual predictions are returned (as a matrix with each column representing predictions from a specific model).}

\item{verbose}{Set to \code{TRUE} to print messages on status and information to the console. Turn this on by default using \code{options(growthcurveSL.verbose=TRUE)}.}

\item{valid_data}{Subject-specific holdout (validation) data for which predictions should be obtained.}
}
\value{
...
}
\description{
Predict for holdout observations only
}

