% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthcurve_model.R
\name{predict_holdoutSL}
\alias{predict_holdoutSL}
\title{Prediction for holdout SuperLearner fit}
\usage{
predict_holdoutSL(modelfit, newdata, add_subject_data = FALSE,
  verbose = getOption("growthcurveSL.verbose"))
}
\arguments{
\item{modelfit}{Model fit object returned by \code{\link{get_fit}} function.}

\item{newdata}{Subject-specific holdout (validation) data for which predictions should be obtained.}

\item{add_subject_data}{Set to \code{TRUE} to add the subject-level data to the resulting predictions (returned as a data.table).
When \code{FALSE} (default) only the actual predictions are returned (as a matrix with each column representing predictions from a specific model).}

\item{verbose}{Set to \code{TRUE} to print messages on status and information to the console. Turn this on by default using \code{options(growthcurveSL.verbose=TRUE)}.}
}
\value{
...
}
\description{
Prediction for holdout SuperLearner fit
}

