% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelingSL_main.R
\name{predict_model}
\alias{predict_model}
\title{Predict for new dataset}
\usage{
predict_model(modelfit, newdata, predict_only_bestK_models,
  verbose = getOption("growthcurveSL.verbose"))
}
\arguments{
\item{modelfit}{Model fit object returned by \code{\link{fit_model}} function.}

\item{newdata}{...}

\item{predict_only_bestK_models}{... Not implemented ...}

\item{verbose}{Set to \code{TRUE} to print messages on status and information to the console. Turn this on by default using \code{options(growthcurveSL.verbose=TRUE)}.}
}
\value{
...
}
\description{
Predict for new dataset
}

