% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthcurve_features.R
\name{predict_save_tgrid}
\alias{predict_save_tgrid}
\title{Predict and save the entire growth curve (grid of time points).}
\usage{
predict_save_tgrid(SLfit, data, ID, t_name, y, tmin = 1, tmax = 500,
  incr = 2, hold_column = NULL, file.name = NULL,
  file.path = getOption("growthcurveSL.file.path"))
}
\arguments{
\item{SLfit}{SuperLearner fit returned by \code{\link{fit_holdoutSL}} or  \code{\link{fit_cvSL}}.}

\item{data}{Input data used for model training.}

\item{ID}{A character string name of the column that contains the unique subject identifiers.}

\item{t_name}{A character string name of the column with integer-valued measurement time-points (in days, weeks, months, etc).}

\item{y}{A character string name of the column that represent the response variable in the model.}

\item{tmin}{Min t value of the grid}

\item{tmax}{Max t value of the grid}

\item{incr}{Increment value for the grid of \code{t}'s}

\item{hold_column}{Column name in \code{dataDT} indicating the holdout observations (if used). Leave as \code{NULL} if no \code{hold_column} is present.}

\item{file.name}{A file name (without the .csv extension) in which the prediction dataset will be saved. Leave as NULL is no saving is necessary.}

\item{file.path}{A directory path in which the predictions file should be saved.}
}
\value{
A \code{data.table} with subject IDs, grid of equally spaced time-points and the corresponding growth curve predictions.
The relevant subject summaries and predictors used for training will be also included in the output data.
In addition, the output dataset also contains an indicator column 'train_point', set to \code{TRUE} for all (ID,time-points) that also appear
in the input data \code{dataDT}. That is 'train_point' indicates if the row might have been previously used for model training.
Finally, if the argument 'hold_column' is not NULL, the output dataset will contain the indicator column of holdout observations
(named according to hold_column argument).
}
\description{
A thin wrapper for \code{define_features_drop}, \code{define_tgrid} and \code{predict_SL} functions.
Evaluates predictions from the existing SuperLearner fit using an entire grid of time points.
Optionally, when \code{file.name} is not \code{NULL}, the resulting dataset of predictions is saved as a csv file.
}

