% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SuperLearner_main.R
\name{predict_SL}
\alias{predict_SL}
\title{Predict for holdout or CV SuperLearner fits}
\usage{
predict_SL(modelfit, newdata, add_subject_data = FALSE, subset_idx = NULL,
  holdout = FALSE, force_data.table = TRUE,
  verbose = getOption("longGriDiSL.verbose"))
}
\arguments{
\item{modelfit}{Model fit object returned by \code{\link{fit_holdoutSL}} or  \code{\link{fit_cvSL}}.}

\item{newdata}{Subject-specific data for which predictions should be obtained.
If missing then the predictions for the training data will be typically returned.
See \code{holdout} for discussion of alternative cases.}

\item{add_subject_data}{Set to \code{TRUE} to add the subject-level data to the resulting predictions (returned as a data.table).
When \code{FALSE} (default) only the actual predictions are returned (as a matrix with each column representing predictions from a specific model).}

\item{subset_idx}{A vector of row indices in \code{newdata} for which the predictions should be obtain.
Default is \code{NULL} in which case all observations in \code{newdata} will be used for prediction.}

\item{holdout}{Set to \code{TRUE} for out-of-sample predictions for validation folds (out-of-sample observations) or holdouts.
When \code{newdata} is missing there are two possible types of holdout predictions, depending on the modeling approach.
1. For \code{\link{fit_holdoutSL}} the default holdout predictions will be based on validation data.
2. For \code{\link{fit_cvSL}} the default is to leave use the previous out-of-sample (holdout) predictions from the training data.}

\item{force_data.table}{...}

\item{verbose}{Set to \code{TRUE} to print messages on status and information to the console. Turn this on by default using \code{options(longGriDiSL.verbose=TRUE)}.}
}
\value{
A data.table of subject level predictions (subject are rows, columns are different models)
or a data.table with subject level covariates added along with model-based predictions.
}
\description{
Predict for holdout or CV SuperLearner fits
}

