% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_estimation.R
\name{predictCurve}
\alias{predictCurve}
\title{Predict for a holdout set}
\usage{
predictCurve(OData, tmin, tmax, verbose = getOption("growthcurveSL.verbose"))
}
\arguments{
\item{OData}{Input data object created by \code{importData} function.}

\item{tmin}{...}

\item{tmax}{...}

\item{verbose}{Set to \code{TRUE} to print messages on status and information to the console. Turn this on by default using \code{options(growthcurveSL.verbose=TRUE)}.}
}
\value{
...
}
\description{
Predict for a holdout set
}

