% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_estimation.R
\name{predictCV}
\alias{predictCV}
\title{Predict for cross-validation sets}
\usage{
predictCV(modelfit, data_valid, modelIDs,
  verbose = getOption("growthcurveSL.verbose"))
}
\arguments{
\item{modelfit}{Model fit object returned by \code{\link{get_fit}} function.}

\item{modelIDs}{... Not implemented ...}

\item{verbose}{Set to \code{TRUE} to print messages on status and information to the console. Turn this on by default using \code{options(growthcurveSL.verbose=TRUE)}.}

\item{OData}{Input data object created by \code{importData} function.}

\item{OData_valid}{...}
}
\value{
...
}
\description{
Predict for cross-validation sets
}

