% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{prepare_data}
\alias{prepare_data}
\title{Wrapper for several data processing functions.}
\usage{
prepare_data(data, OUTCOME, vars_to_numeric, vars_to_int, skip_vars,
  drop_vars = FALSE, verbose = FALSE)
}
\arguments{
\item{data}{Input dataset, can be a \code{data.frame} or a \code{data.table}.}

\item{OUTCOME}{Character name of the column of outcomes.}

\item{vars_to_numeric}{Column name(s) that should be converted to numeric type.}

\item{vars_to_int}{Column name(s) that should be converted to integer type.}

\item{skip_vars}{These columns will not be converted into other types}

\item{drop_vars}{Drop the variables in skip_vars and drop the factor variables that were re-coded into binary dummies from the resulting analytic dataset?}

\item{verbose}{Print processing info}
}
\description{
Clean up the input data by dropping observations with missing outcomes \code{OUTCOME},
convert desired columns into numerics (\code{vars_to_numeric}),
convert all logical columns into binary integers
convert all character columns into factors
convert all factors into integers,
by defining additional dummy variables for every factor with > 2 levels.
}
