% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelPredictionModel.R
\name{print.PredictionModel}
\alias{print.PredictionModel}
\title{S3 methods for printing model fit summary for PredictionModel R6 class object}
\usage{
\method{print}{PredictionModel}(x, model_stats = FALSE, all_fits = FALSE,
  ...)
}
\arguments{
\item{x}{The model fit object produced by functions \code{get_fit}.}

\item{model_stats}{Also print some model summaries?}

\item{all_fits}{Print all of the modeling fits contained in this object? Warning: this may produce a lot of output!}

\item{...}{Additional options passed on to \code{print.PredictionModel}.}
}
\description{
Prints the modeling summaries
}
