% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelPredictionStack.R
\name{print.PredictionStack}
\alias{print.PredictionStack}
\title{S3 methods for printing model fit summary for PredictionModel R6 class object}
\usage{
\method{print}{PredictionStack}(x, ...)
}
\arguments{
\item{x}{The model fit object produced by functions \code{make_PredictionStack}.}

\item{...}{Additional options passed on to \code{print.PredictionModel}.}
}
\description{
Prints the modeling summaries
}
