% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{print_GriDiSL_opts}
\alias{print_GriDiSL_opts}
\title{Print Current Option Settings for \code{GriDiSL}}
\usage{
print_GriDiSL_opts()
}
\value{
Invisibly returns a list of \code{GriDiSL} options.
}
\description{
Print Current Option Settings for \code{GriDiSL}
}
\seealso{
\code{\link{set_all_GriDiSL_options}}
}

