% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{print_growthcurveSL_opts}
\alias{print_growthcurveSL_opts}
\title{Print Current Option Settings for \code{growthcurveSL}}
\usage{
print_growthcurveSL_opts()
}
\value{
Invisibly returns a list of \code{growthcurveSL} options.
}
\description{
Print Current Option Settings for \code{growthcurveSL}
}
\seealso{
\code{\link{set_all_growthcurveSL_options}}
}

