% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{print_longDiSL_opts}
\alias{print_longDiSL_opts}
\title{Print Current Option Settings for \code{longDiSL}}
\usage{
print_longDiSL_opts()
}
\value{
Invisibly returns a list of \code{longDiSL} options.
}
\description{
Print Current Option Settings for \code{longDiSL}
}
\seealso{
\code{\link{set_all_longDiSL_options}}
}

