% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{print_longGriDiSL_opts}
\alias{print_longGriDiSL_opts}
\title{Print Current Option Settings for \code{longGriDiSL}}
\usage{
print_longGriDiSL_opts()
}
\value{
Invisibly returns a list of \code{longGriDiSL} options.
}
\description{
Print Current Option Settings for \code{longGriDiSL}
}
\seealso{
\code{\link{set_all_longGriDiSL_options}}
}

