% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryS3methods.R
\name{print_tables.brokenstick}
\alias{print_tables.brokenstick}
\title{S3 methods for printing model fit summaries for brokenstick class}
\usage{
\method{print_tables}{brokenstick}(model, only.coefs = FALSE, ...)
}
\description{
S3 methods for printing model fit summaries for brokenstick class
}

