% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegressionControlClasses.R
\docType{class}
\name{RegressionClass}
\alias{RegressionClass}
\title{R6 class that defines a learner / model for fitting E(Y|W)}
\format{An \code{\link{R6Class}} generator object}
\usage{
RegressionClass
}
\description{
This R6 class defines fields and methods that controls all the parameters of a regression model / machine learning algorithm.
}
\details{
\itemize{
\item{\code{outvar.class}} - Character vector indicating a class of each outcome var: \code{bin} / \code{cont} / \code{cat}.
\item{\code{outvar}} - Character vector of regression outcome variable names.
\item{\code{predvars}} - Predictors (names).
\item{\code{subset_vars}} - Subset variables (later evaluated to logical vector based on non-missing (!is.na()) values of these variables).
\item{\code{subset_exprs}} - Subset expressions (later evaluated to logical vector in the envir of the data).
\item{\code{ReplMisVal0}} - Logical, if TRUE all gvars$misval among predicators are replaced with with gvars$misXreplace (0).
}
}
\section{Methods}{

\describe{
  \item{\code{new(...)}}{Uses the arguments to instantiate an object of R6 class and define the future regression model.}
  \item{\code{ChangeManyToOneRegresssion(k_i, reg)}}{ Take a clone of a parent \code{RegressionClass} (\code{reg}) for \code{length(self$outvar)} regressions
   and set self to a single univariate \code{k_i} regression for outcome \code{self$outvar[[k_i]]}.}
  \item{\code{ChangeOneToManyRegresssions(regs_list)}}{ Take the clone of a parent \code{RegressionClass} for univariate (continuous outvar) regression
    and set self to \code{length(regs_list)} bin indicator outcome regressions.}
  \item{\code{resetS3class()}}{...}
}
}

\section{Active Bindings}{

\describe{
  \item{\code{S3class}}{...}
  \item{\code{get.reg}}{...}
}
}
\keyword{R6}
\keyword{class}

