% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthcurve_model.R
\name{score_CVgrowthcurveSL}
\alias{score_CVgrowthcurveSL}
\title{Predict for cross-validation sets}
\usage{
score_CVgrowthcurveSL(modelfit, valid_data,
  verbose = getOption("growthcurveSL.verbose"))
}
\arguments{
\item{modelfit}{Model fit object returned by \code{\link{get_fit}} function.}

\item{valid_data}{New validation data for external CV}

\item{verbose}{Set to \code{TRUE} to print messages on status and information to the console. Turn this on by default using \code{options(growthcurveSL.verbose=TRUE)}.}
}
\value{
new CV-based MSEs
}
\description{
Predict for cross-validation sets
}

