% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{set_all_growthcurveSL_options}
\alias{set_all_growthcurveSL_options}
\title{Setting \code{growthcurveSL} Options}
\usage{
set_all_growthcurveSL_options(fit.package = c("h2o", "speedglm", "glm",
  "brokenstick", "face"), fit.algorithm = c("glm", "gbm", "randomForest",
  "deeplearning", "GridLearner"), maxncats = 20)
}
\arguments{
\item{fit.package}{Specify the default package for performing model fitting: c("speedglm", "glm", "h2o")}

\item{fit.algorithm}{Specify the default fitting algorithm: c("glm", "gbm", "randomForest", "SuperLearner")}

\item{maxncats}{Max number of unique categories a categorical variable can have. More than these number and it is deemed continuous.}
}
\value{
Invisibly returns a list with old option settings.
}
\description{
Options that control \code{growthcurveSL} package.
\strong{Will reset all unspecified options (omitted arguments) to their default values}.
The preferred way to set options for \code{growthcurveSL} is to use \code{\link{growthcurveSLOptions}}, which allows specifying individual options without having to reset all other options.
To reset all options to their defaults simply run \code{set_all_growthcurveSL_options()} without any parameters/arguments.
}
\seealso{
\code{\link{growthcurveSLOptions}}, \code{\link{print_growthcurveSL_opts}}
}

