% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryS3methods.R
\name{summary.xgb.Booster}
\alias{summary.xgb.Booster}
\alias{summary.xgb.cv.synchronous}
\title{S3 methods for fit summary from xgboost}
\usage{
\method{summary}{xgb.Booster}(xgb.model, ...)

\method{summary}{xgb.cv.synchronous}(xgb.model, ...)
}
\arguments{
\item{xgb.model}{The model fit object produced by xgboost (and extracted with \code{getmodel_byname}).}

\item{...}{Additional options (not used)}
}
\value{
The markdown-formated model summary returned by \code{pander::pander_return}.
}
\description{
Prints the modeling summary for the xgboost model fit (see \code{xgboost} R package).
}

