% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_funs.R
\name{to_int}
\alias{to_int}
\title{Convert specific columns in \code{vars} to integers}
\usage{
to_int(data, vars)
}
\arguments{
\item{data}{Input dataset, as \code{data.table}.}

\item{vars}{Column name(s) that should be converted to numeric type.}
}
\description{
Convert specific columns in \code{vars} to integers
}

