% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_funs.R
\name{add_CVfolds_ind}
\alias{add_CVfolds_ind}
\title{Define a column of fold indicators for V-fold cross-validation}
\usage{
add_CVfolds_ind(data, ID, nfolds = 5, fold_column = "fold", seed = NULL)
}
\arguments{
\item{data}{Input dataset, can be a \code{data.frame} or a \code{data.table}.}

\item{ID}{A character string name of the column that contains the unique subject identifiers.}

\item{nfolds}{Number of unique folds (same fold is always assigned to all observations that share the same ID).}

\item{fold_column}{A name of the column that will contain the fold indicators}

\item{seed}{Random number seed for selecting a random fold.}
}
\value{
An input data with added fold indicator column (as ordered factor with levels 1:nfolds).
}
\description{
The input data is assumed to have repeated observations per subjects, the folds are defined as clustered by subject IDs.
}
