% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hbgd_fittedTrajectory.R
\name{create_fittedTrajectory_1subj}
\alias{create_fittedTrajectory_1subj}
\title{Create a fittedTrajectory object with fits for a single subject}
\usage{
create_fittedTrajectory_1subj(subj_dat_1subj, grid_fits_dat_1subj,
  holdout_fits_dat_1subj = NULL, t_var, y_var, sex_var, method,
  xy_pair_name = c("agedays", "htcm"), grid_fits_var = "SL.preds",
  holdout_fits_var = "holdout_fit", checkpoints = c(50, 100),
  fun_y_to_raw = hbgd::who_zscore2htcm, fun_y_to_z = function(x, y, ...)
  return(y))
}
\arguments{
\item{subj_dat_1subj}{Data on a single subject}

\item{grid_fits_dat_1subj}{Grid of equally spaced model predictions for a single subject}

\item{holdout_fits_dat_1subj}{Data with model predictions for holdout datapoints for a single subject (optional)}

\item{t_var}{A character string name of the column with integer-valued measurement time-points (in days, weeks, months, etc).}

\item{y_var}{A character string name of the column that represent the response variable in the model.}

\item{sex_var}{A character string name identifying the subject's gender in \code{subj_dat_1subj}.}

\item{method}{A fitting method.}

\item{xy_pair_name}{...}

\item{grid_fits_var}{...}

\item{holdout_fits_var}{...}

\item{checkpoints}{...}

\item{fun_y_to_raw}{...}

\item{fun_y_to_z}{...}
}
\value{
An object of class \code{fittedTrajectory} for further downstream analysis.
}
\description{
Creates an object of class \code{fittedTrajectory} with growth curve predictions.
This can be used for further growth curve analysis with hbgd R package.
Requires the the dataset containing the predicted growth curves for a single subject \code{grid_fits_dat_1subj}.
}

