% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_main.R
\name{defLearner}
\alias{+.ModelStack}
\alias{defGrid}
\alias{defLearner}
\title{Interface for defining models}
\usage{
defLearner(estimator, x, ...)

defGrid(estimator, x, search_criteria, param_grid, ...)

\method{+}{ModelStack}(learner1, learner2)
}
\arguments{
\item{estimator}{A character string name of package and estimator (algorithm) name, separated by "__".}

\item{x}{A vector containing the subset of the names of the predictor variables to use in building this
particular learner or grid. This argument can be used to over-ride the values of \code{x} provided to \code{fit} function.
As such, the names supplied here must always be a subset of the names specified to \code{fit}.
When this argument is missing (default) the column names provided to \code{fit} are used as predictors in building this model / grid.}

\item{...}{Additional modeling parameters to be passed on directly to the modeling function.}

\item{search_criteria}{Search criteria}

\item{param_grid}{Grid of modeling parameters}

\item{learner1}{An object returned by a call to \code{defLearner} or \code{defGrid} functions.}

\item{learner2}{An object returned by a call to \code{defLearner} or \code{defGrid} functions.}
}
\description{
Use \code{defLearner} to define a single model and \code{defGrid} to define a grid of multple models.
}

