% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_main.R
\name{defGrid}
\alias{defGrid}
\title{Define a grid of modeling algorithm(s), package and parameters}
\usage{
defGrid(estimator, search_criteria, param_grid, ...)
}
\arguments{
\item{estimator}{A character string name of package and estimator (algorithm) name, separated by "_".}

\item{search_criteria}{Search criteria}

\item{param_grid}{Grid of modeling parameters}

\item{...}{Additional modeling parameters to be passed to modeling function.}
}
\description{
Define a grid of modeling algorithm(s), package and parameters
}

