% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthcurve_features.R
\name{define_tgrid}
\alias{define_tgrid}
\title{Define dataset of time-point grids for prediction}
\usage{
define_tgrid(dataDT, ID, t_name, y, tmin = 1, tmax = 500, incr = 2)
}
\arguments{
\item{dataDT}{Input data with all relevant summaries defined.}

\item{ID}{A character string name of the column that contains the unique subject identifiers.}

\item{t_name}{A character string name of the column with integer-valued measurement time-points (in days, weeks, months, etc).}

\item{y}{A character string name of the column that represent the response variable in the model.}

\item{tmin}{Min t value of the grid}

\item{tmax}{Max t value of the grid}

\item{incr}{Increment value for the grid of t's}
}
\value{
...
}
\description{
Define dataset of time-point grids for prediction
}

