% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_main.R
\name{fit}
\alias{fit}
\alias{fit.ModelStack}
\title{Fit Discrete SuperLearner (Ensemble)}
\usage{
fit(...)

\method{fit}{ModelStack}(models, method = c("none", "cv", "holdout"), data,
  ID, t_name, x, y, nfolds = NULL, fold_column = NULL, hold_column = NULL,
  hold_random = FALSE, seed = NULL, refit = TRUE,
  verbose = getOption("GriDiSL.verbose"), ...)
}
\arguments{
\item{...}{Additional arguments that will be passed on to \code{fit_model} function.}

\item{models}{...}

\item{method}{The type of model selection procedure when fitting several models at once. Possible options are "none", "cv", and "holdout".}

\item{data}{Input dataset, can be a \code{data.frame} or a \code{data.table}.}

\item{ID}{A character string name of the column that contains the unique subject identifiers.}

\item{t_name}{A character string name of the column with integer-valued measurement time-points (in days, weeks, months, etc).}

\item{x}{A vector containing the names of predictor variables to use for modeling. If x is missing, then all columns except \code{ID}, \code{y} are used.}

\item{y}{A character string name of the column that represent the response variable in the model.}

\item{nfolds}{Number of folds to use in cross-validation.}

\item{fold_column}{The name of the column in the input data that contains the cross-validation fold indicators (must be an ordered factor).}

\item{hold_column}{The name of the column that contains the holdout observation indicators (TRUE/FALSE) in the input data.
This holdout column must be defined and added to the input data prior to calling this function.}

\item{hold_random}{Logical, specifying if the holdout observations should be selected at random.
If FALSE then the last observation for each subject is selected as a holdout.}

\item{seed}{Random number seed for selecting a random holdout.}

\item{refit}{Set to \code{TRUE} (default) to refit the best estimator using the entire dataset.
When \code{FALSE}, it might be impossible to make predictions from this model fit.}

\item{verbose}{Set to \code{TRUE} to print messages on status and information to the console. Turn this on by default using \code{options(GriDiSL.verbose=TRUE)}.}

\item{params}{Parameters specifying the type of modeling procedure to be used.}
}
\value{
...
}
\description{
Define and fit discrete SuperLearner for growth curve modeling.
Model selection (scoring) is based on MSE for a single random (or last) holdout data-point for each subject.
This is in contrast to the model selection with V-fold cross-validated MSE in \code{\link{fit_cvSL}},
which leaves the entire subjects (entire growth curves) outside of the training sample.
}

