% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_estimation.R
\name{fit_model}
\alias{fit_model}
\title{Generic modeling function for any longitudinal data.}
\usage{
fit_model(ID, t_name, x, y, data, data_validation, params,
  verbose = getOption("growthcurveSL.verbose"))
}
\arguments{
\item{ID}{A character string name of the column that contains the unique subject identifiers.}

\item{t_name}{A character string name of the column with integer-valued measurement time-points (in days, weeks, months, etc).}

\item{x}{A vector containing the names of predictor variables to use for modeling. If x is missing, then all columns except \code{ID}, \code{y} are used.}

\item{y}{A character string name of the column that represent the response variable in the model.}

\item{data}{Input dataset, can be a \code{data.frame} or a \code{data.table}.}

\item{data_validation}{Optional \code{data.frame} or \code{data.table} with validation data. When provided, this dataset will be used for scoring the model fit(s).}

\item{params}{Parameters specifying the type of modeling procedure to be used.}

\item{verbose}{Set to \code{TRUE} to print messages on status and information to the console. Turn this on by default using \code{options(growthcurveSL.verbose=TRUE)}.}
}
\value{
...
}
\description{
Generic modeling function for any longitudinal data.
}

