% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_SuperLearner_funs.R
\name{get_train_data}
\alias{get_train_data}
\title{Get training data used by the modeling object}
\usage{
get_train_data(modelfit)
}
\arguments{
\item{modelfit}{A model object of class \code{PredictionModel} returned by function \code{fit_model} or \code{fit}.}
}
\value{
\code{data.table} that was used for model training.
}
\description{
Wrapper function for obtaining the training dataset saved in the modeling object.
}
