% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{GriDiSLOptions}
\alias{GriDiSLOptions}
\title{Querying/setting a single \code{GriDiSL} option}
\usage{
GriDiSLOptions(o, value)
}
\arguments{
\item{o}{Option name (string). See \code{\link{set_all_GriDiSL_options}}.}

\item{value}{Value to assign (optional)}
}
\description{
To list all \code{GriDiSL} options, just run this function without any parameters provided. To query only one value, pass the first parameter. To set that, use the \code{value} parameter too.
}
\details{
The arguments of \code{\link{set_all_GriDiSL_options}} list all available \code{GriDiSL} options.
}
\examples{
\dontrun{
GriDiSLOptions()
GriDiSLOptions('fit.package')
GriDiSLOptions('fit.package', 'h2o')
}
}
\seealso{
\code{\link{set_all_GriDiSL_options}}
}

