% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{gridislOptions}
\alias{gridislOptions}
\title{Querying/setting a single \code{gridisl} option}
\usage{
gridislOptions(o, value)
}
\arguments{
\item{o}{Option name (string).}

\item{value}{Value to assign (optional)}
}
\description{
To list all \code{gridisl} options, just run this function without any parameters provided. To query only one value, pass the first parameter.
To set that, use the \code{value} parameter too.
}
\examples{
\dontrun{
gridislOptions()
gridislOptions('fit.package')
gridislOptions('fit.package', 'h2o')
}
}
