% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_funs.R
\name{importData}
\alias{importData}
\title{Import data, define nodes (columns) and define input data R6 object}
\usage{
importData(data, ID = "Subject_ID", t_name = "time_period", covars,
  OUTCOME = "Y", verbose = getOption("gridisl.verbose"))
}
\arguments{
\item{data}{Input dataset, can be a \code{data.frame} or a \code{data.table}.}

\item{ID}{A character string name of the column that contains the unique subject identifiers.}

\item{t_name}{A character string name of the column with integer-valued measurement time-points (in days, weeks, months, etc).}

\item{covars}{Names of predictors (covariates) in the data.}

\item{OUTCOME}{Character name of the column containing outcomes.}

\item{verbose}{Set to \code{TRUE} to print messages on status and information to the console. Turn this on by default using \code{options(gridisl.verbose=TRUE)}.}
}
\value{
An R6 object that contains the input data. This can be passed as an argument to \code{get_fit} function.
}
\description{
Import data, define nodes (columns) and define input data R6 object
}
