% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_funs.R
\name{plotMSEs}
\alias{plotMSEs}
\title{Plot the top (smallest) validation MSEs for an ensemble of prediction models}
\usage{
plotMSEs(PredictionModel, K = 1, interactive = FALSE)
}
\arguments{
\item{PredictionModel}{Must be an R6 object of class \code{PredictionModel} (returned by \code{get_fit} function)
or an object of class \code{PredictionStack} (returned by \code{make_PredictionStack} function).
Must also contain validation /test set predictions and corresponding MSEs.}

\item{K}{How many top (smallest) MSEs should be plotted? Default is 5.}

\item{interactive}{Setting this to \code{TRUE} will produce an interactive plot in html format using the package \code{ggiraph}.}
}
\description{
Plot the top (smallest) validation MSEs for an ensemble of prediction models
}
