% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_SuperLearner_funs.R
\name{predict_generic}
\alias{predict_generic}
\title{Generic SuperLearner prediction function}
\usage{
predict_generic(modelfit, newdata, add_subject_data = FALSE,
  subset_idx = NULL, best_only = TRUE, holdout = FALSE,
  force_data.table = TRUE, verbose = getOption("gridisl.verbose"))
}
\arguments{
\item{modelfit}{Model fit object returned by \code{\link{fit}} functions. Must be an object of class \code{PredictionModel} or \code{PredictionStack}.}

\item{newdata}{Subject-specific data for which predictions should be obtained.}

\item{add_subject_data}{Set to \code{TRUE} to add the subject-level data to the resulting predictions (returned as a data.table).
When \code{FALSE} (default) only the actual predictions are returned (as a matrix with each column representing predictions from a specific model).}

\item{subset_idx}{A vector of row indices in \code{newdata} for which the predictions should be obtain.
Default is \code{NULL} in which case all observations in \code{newdata} will be used for prediction.}

\item{best_only}{Set to \code{TRUE} (default) to obtain predictions from the top-ranked model (based on validation or CV MSE).
When \code{FALSE} the attempt will to made to obtain predictions from all models.
Note that when \code{holdout} is \code{FALSE} and \code{best_only} is \code{TRUE},
the predictions will be based on the best scoring model that was re-trained on all available data.}

\item{holdout}{Set to \code{TRUE} for out-of-sample predictions for validation folds or holdouts.}

\item{force_data.table}{Force the prediction result to be \code{data.table}.}

\item{verbose}{Set to \code{TRUE} to print messages on status and information to the console.
Turn this on by default using \code{options(gridisl.verbose=TRUE)}.}
}
\value{
A data.table of subject level predictions (subject are rows, columns are different models)
or a data.table with subject level covariates added along with model-based predictions.
}
\description{
Generic SuperLearner prediction function
}
