% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelPredictionModel.R
\docType{class}
\name{PredictionModel}
\alias{PredictionModel}
\title{R6 class for fitting and making predictions for a single or a grid of regression models E(outvar | predvars)}
\format{An \code{\link{R6Class}} generator object}
\usage{
PredictionModel
}
\description{
This R6 class can request, store and manage the design matrix Xmat, as well as the outcome Y
 The class has methods that perform queries to data storage R6 class DataStorageClass to get appropriate data columns & row subsets
}
\details{
\itemize{
\item{GLMpackage} - Controls which package will be used for performing model fits (\code{glm} or \code{speedglm}).
\item{ModelFitObject} - Pointer to an instance of \code{ModelFitObject} class that contains the data.
}
}
\section{Methods}{

\describe{
  \item{\code{new(reg)}}{Uses \code{reg} R6 \code{\link{RegressionClass}} object to instantiate a new model for a
  logistic regression with binary outcome.}
  \item{\code{show()}}{Print information on outcome and predictor names used in this regression model}
  \item{\code{fit()}}{...}
  \item{\code{copy.fit()}}{...}
  \item{\code{predict()}}{...}
  \item{\code{copy.predict()}}{...}
  \item{\code{predictAeqa()}}{...}
}
}
\keyword{R6}
\keyword{class}

