% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{print_gridisl_opts}
\alias{print_gridisl_opts}
\title{Print Current Option Settings for \code{gridisl}}
\usage{
print_gridisl_opts()
}
\value{
Invisibly returns a list of \code{gridisl} options.
}
\description{
Print Current Option Settings for \code{gridisl}
}
\seealso{
\code{\link{set_all_gridisl_options}}
}
