% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_funs.R
\name{save_best_model}
\alias{save_best_model}
\title{Save the best performing h2o model}
\usage{
save_best_model(modelfit, file.path = getOption("gridisl.file.path"))
}
\arguments{
\item{modelfit}{A model object of class \code{PredictionModel} returned by functions \code{fit_model} or \code{fit}.}

\item{file.path}{Specify the directory where the model object file should be saved.}
}
\description{
Save the best performing h2o model
}
