% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{set_all_gridisl_options}
\alias{set_all_gridisl_options}
\title{Setting \code{gridisl} Options}
\usage{
set_all_gridisl_options(fit.package = c("speedglm", "glm", "h2o", "xgboost",
  "brokenstick", "face"), fit.algorithm = c("glm", "gbm", "randomForest",
  "drf", "deeplearning", "grid", "resid_grid"))
}
\arguments{
\item{fit.package}{Specify the default package for performing model fitting: c("speedglm", "glm", "h2o")}

\item{fit.algorithm}{Specify the default fitting algorithm: c("glm", "gbm", "randomForest", "SuperLearner")}
}
\value{
Invisibly returns a list with old option settings.
}
\description{
Options that control \code{gridisl} package.
\strong{Will reset all unspecified options (omitted arguments) to their default values}.
The preferred way to set options for \code{gridisl} is to use \code{\link{gridislOptions}}, which allows specifying individual options without having to reset all other options.
To reset all options to their defaults simply run \code{set_all_gridisl_options()} without any parameters/arguments.
}
\seealso{
\code{\link{gridislOptions}}, \code{\link{print_gridisl_opts}}
}
